Fixed [Made index long, better practice]

+++ 0x1A-hash_tables/5-hash_table_print.c
@@ -9,7 +9,7 @@
 void hash_table_print(const hash_table_t *ht)
 {
 
-	unsigned int i, j;
+	unsigned long int i, j;
 	hash_node_t *help;
 
 	if (!ht)
